Refactor wish detail page
