+++ src/pages/wishDetail/components/WishInformation/index.js
@@ -1,8 +1,12 @@
 import React from 'react';
 import media from '@kiwicom/orbit-components/lib/utils/mediaQuery';
 import styled, { css } from 'styled-components';
+
+// components
 import { Badge, Stack, Text, Heading } from '@kiwicom/orbit-components/lib';
-import WishInformationHeader from '../../postDetails/PostDetailsHeader';
+import WishInformationHeader from '@components/postDetails/PostDetailsHeader';
+
+// constants and utils
 import { wishes } from '@constants/postType';
 
 const WishInformationBodyContainer = styled.div`
