+++ src/pages/wishDetail/components/index.js
@@ -1,0 +1,2 @@
+export { default as Map } from './Map';
+export { default as WishInformation } from './WishInformation';
